added jwt auth

+++ src/main/java/com/wipro/app/controller/UserController.java
@@ -1,6 +1,8 @@
 package com.wipro.app.controller;
 
+import com.wipro.app.model.Role;
 import com.wipro.app.model.User;
+import com.wipro.app.payload.ApiRequest;
 import com.wipro.app.payload.ApiResponse;
 import com.wipro.app.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,16 +21,40 @@ public class UserController {
     private UserService userService;
 
     @PutMapping("/user/{username}")
-    public ResponseEntity<User> updateUser(@RequestBody User user , @PathVariable String username){
+    public ResponseEntity<?> updateUser(@RequestBody User user , @PathVariable String username){
+        if(this.userService.isUserExist(user.getUsername())){
+            return new ResponseEntity<>(new ApiResponse("username already exists." , false , 400) , HttpStatus.BAD_REQUEST);
+        }
         User updateUser = this.userService.updateUser(user , username);
+        return new ResponseEntity<>(updateUser , HttpStatus.OK);
+    }
+
+    @PutMapping("/user/{username}/role")
+    public ResponseEntity<?> updateRole(@RequestBody ApiRequest<Role> request , @PathVariable String username){
+        if(!this.userService.isUserExist(username)){
+            return new ResponseEntity<>(new ApiResponse("username is not present." , true , 400) , HttpStatus.BAD_REQUEST);
+        }
+        User updateUser = this.userService.updateRole(username , request.getData());
+        return new ResponseEntity<>(updateUser , HttpStatus.OK);
+    }
+
+    @PutMapping("/user/{username}/password")
+    public ResponseEntity<?> updatePassword(@RequestBody ApiRequest<String> request , @PathVariable String username){
+        if(!this.userService.isUserExist(username)){
+            return new ResponseEntity<>(new ApiResponse("username is not present." , true , 400) , HttpStatus.BAD_REQUEST);
+        }
+        User updateUser = this.userService.updatePassword(username , request.getData());
         return new ResponseEntity<>(updateUser , HttpStatus.OK);
     }
 
     @DeleteMapping("/user/{username}")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<ApiResponse> deleteUser(@PathVariable String username){
+        if(!this.userService.isUserExist(username)){
+            return new ResponseEntity<>(new ApiResponse("username is not present." , true , 400) , HttpStatus.BAD_REQUEST);
+        }
         this.userService.deleteUser(username);
-        return new ResponseEntity<>(new ApiResponse("User Deleted Successfully" , true , 200) , HttpStatus.OK);
+        return new ResponseEntity<>(new ApiResponse("user deleted successfully" , true , 200) , HttpStatus.OK);
     }
 
     @GetMapping("/users")
@@ -36,8 +62,12 @@ public class UserController {
         return ResponseEntity.ok(this.userService.getAllUsers());
     }
 
+    //TODO: select query is calling multiple times - FIX ME
     @GetMapping("/user/{username}")
-    public ResponseEntity<User> getUserByUsername(@PathVariable String username){
+    public ResponseEntity<?> getUserByUsername(@PathVariable String username){
+        if(!this.userService.isUserExist(username)){
+            return new ResponseEntity<>(new ApiResponse("username is not present." , true , 400) , HttpStatus.BAD_REQUEST);
+        }
         return ResponseEntity.ok(this.userService.getUserByUsername(username));
     }
 }
